Use MSG_BASE_POS_ECEF to determine base station location (#60)
Add a new BASEPOS option for sbp2rinex. If this option is specified, then the MSG_BASE_POS_ECEF SBP message will be used to determine the base station location. The first position will be written into the header of the RINEX file. Subsequent changes will be stored in the observation file as "new site occupation" events